Matplotlib visulization added.

+++ run_evaluation.sh
@@ -32,19 +32,27 @@ function RUN_COMMAND(){
    echo "./run_champsim_nosuffix.sh ${BINARY} ${N_WARM} ${N_SIM} ${TRACE}"
 }
 
+function STATS_NAME(){
+   local BINARY=$1
+   local TRACE=$2
+   echo "${BINARY}.${TRACE}.${N_SIM}.${N_WARM}.stats"
+}
+
 TRACE_DIR=$PWD/dpc3_traces
 
 mkdir -p ${TRACE_DIR}
 
 EVAL_ROOT_DIR=$PWD/eval
 EVAL_SOURCE_DIR=$PWD/eval/prefetcher
-EVAL_STATS_DIR=$PWD/eval/stats
+EVAL_STATS_DIR=$PWD/eval/stats/not-formatted
 EVAL_STATS_FORMATTED_DIR=$PWD/eval/stats/formatted
+EVAL_SUMM_DIR=$PWD/eval/summary
 
 mkdir -p ${EVAL_ROOT_DIR}
 mkdir -p ${EVAL_SOURCE_DIR}
 mkdir -p ${EVAL_STATS_DIR}
 mkdir -p ${EVAL_STATS_FORMATTED_DIR}
+mkdir -p ${EVAL_SUMM_DIR}
 
 TRACE_TYPE=("client" "server" "spec")
 TRACE_NUM=$(ls -1 ${TRACE_DIR} | wc -l)
@@ -68,6 +76,14 @@ do
 	rm -f $PWD/prefetcher/${PREF}.l1i_pref
 done
 
+if [ -f $PWD/bin/$(BINARY_NAME baseline) ] ; then
+	printf "$(BUILD_COMMAND no) ${BOLD}(Skipped)${NORMAL}\n"
+else
+	printf "$(BUILD_COMMAND no)\n"
+	bash $(BUILD_COMMAND no)
+	mv $PWD/bin/$(BINARY_NAME no) $PWD/bin/$(BINARY_NAME baseline)
+fi
+
 function PROGRESS_BAR(){
     local n=$1
     local i=$2
@@ -83,7 +99,7 @@ for BINARY in $PWD/bin/*
 do
 	[ -f "$BINARY" ] || continue
 	BINARY=`basename ${BINARY}`
-	printf "\n${BOLD}${BINARY}...\n${NORMAL}"
+	printf "\n${BOLD}Simulate ${BINARY}...\n${NORMAL}"
     ((COUNT=1))
 	for TRACE in ${TRACE_DIR}/*
 	do
@@ -96,13 +112,14 @@ do
 				RESULT_FORMATTED_DIR=${EVAL_STATS_FORMATTED_DIR}/${BINARY}/${TYPE}
 				mkdir -p ${RESULT_DIR}
 				mkdir -p ${RESULT_FORMATTED_DIR}
-				if [ -f ${RESULT_DIR}/${BINARY}-${TRACE}-${N_SIM}M.stats ] && [ -f ${RESULT_FORMATTED_DIR}/${BINARY}-${TRACE}-${N_SIM}M.stats ] ; then
+				RESULT=$(STATS_NAME ${BINARY} ${TRACE})
+				if [ -f ${RESULT_DIR}/${RESULT} ] && [ -f ${RESULT_FORMATTED_DIR}/${RESULT} ] ; then
 					printf "$(RUN_COMMAND ${BINARY} ${TRACE}.champsimtrace.xz) ${BOLD}(Skipped)${NORMAL}"
 				else
 					printf "$(RUN_COMMAND ${BINARY} ${TRACE}.champsimtrace.xz)"
 					bash $(RUN_COMMAND ${BINARY} ${TRACE}.champsimtrace.xz)
-					cp $PWD/results_${N_SIM}M/${TRACE}.champsimtrace.xz-${BINARY}${OPTION}.txt ${RESULT_DIR}/${BINARY}-${TRACE}-${N_SIM}M.stats
-					cp $PWD/champsim-stats.json ${RESULT_FORMATTED_DIR}/${BINARY}-${TRACE}-${N_SIM}M.stats
+					cp $PWD/results_${N_SIM}M/${TRACE}.champsimtrace.xz-${BINARY}${OPTION}.txt ${RESULT_DIR}/${RESULT}
+					cp $PWD/champsim-stats.json ${RESULT_FORMATTED_DIR}/${RESULT}
 				fi
 			fi
 		done
@@ -116,4 +133,8 @@ done
 printf "\n"
 
 rm -rf $PWD/results_${N_SIM}M
-rm champsim-stats.json
+[ -f champsim-stats.json ] && rm champsim-stats.json
+
+printf "${CYAN}${BOLD}Summarizing all the results...\n${NORMAL}"
+printf "python eval.py\n\n"
+python eval.py
